tasks/task0029.py
"""
The marketing team is spending way too much time typing in hashtags.
Let's help them with our own Hashtag Generator!

Here's the deal:

It must start with a hashtag (#).
All words must have their first letter capitalized.
If the final result is longer than 140 chars it must return false.
If the input or the result is an empty string it must return false.

Examples:

" Hello there thanks for trying my Kata"  =>  "#HelloThereThanksForTryingMyKata"
"    Hello     World   "                  =>  "#HelloWorld"
""                                        =>  false

"""

from typing import Union


def solution(s: str) -> Union[str, bool]:
    if s == "":
        return False
    res = "#" + s.title().replace(" ", "")
    if len(res) > 140:
        return False

    return res
